accept_green_stagings: force accept the identified green ones
We still only accept the stagings that were identified as being ready
upfront, but in case multiple letters were identified as acceptable,
it can easily happen for the later ones no longer to be green at the
time we want to accept it. Mostly happens when packages are removed
from the rings, which causes the scheduler to re-evaluate all stagings.