New tool: staging-impact
Quite often, there are build failures in a staging project that are not immediately
obvious which package in the staging project might be responsible for it.

This tool helps identifying possible candidates (the final debug remains in human hands)

Usage: staging-impact.sh PRJ PKG REP ARCH

Output will be something in the form of (trimmed example)

openSUSE:Factory:Staging:E/rrdtool:
   Mesa
   brotli
…
   libxshmfence
   libyaml
 * pango
   perl-Encode-Locale
…

The interesting info is the '* pango' in this case: all the other packages listed are used
to build rrdtool (identified using osc dependson), but only pango is having an SR staged in
Staging:E; the others are inherited packages without source changes.

Based on this output, one can now deduce pango as the most likely culprit